+++ packages/highperformer/src/workers/categoryCentroids.schemas.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 export const CategoryCentroidsMessageSchema = z.object({
   type: z.literal("categoryCentroids"),
   positions: z.custom<Float32Array>((v) => v instanceof Float32Array),
-  codes: z.custom<Uint8Array>((v) => v instanceof Uint8Array),
+  codes: z.custom<Uint16Array>((v) => v instanceof Uint16Array),
   numCategories: z.number().int().positive(),
 });
